chore: rename scenario file
