+++ src/index.js
@@ -1,5 +1,5 @@
 import { clix } from './clix.js';
 
-export { ScenarioExecutor } from './scenario.js';
+export { ScenarioExecutor } from './scenario-executor.js';
 
 export default clix;
